# src/docs/openapi.yml
openapi: 3.0.0
info:
  title: Rain or Shine API
  version: 1.0.0
  description: |
    API for automatically adding weather data to Strava activities.

    This service integrates with Strava's webhook system to automatically
    add weather information to new activities based on their time and location.
  contact:
    name: API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication using cookies
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-token
      description: Admin authentication token for administrative operations
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
        message:
          type: string
          description: Human-readable message about the operation
        data:
          type: object
          description: Response data (varies by endpoint)
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code (optional)
...
paths:
...
  /api/auth/strava:
    get:
      summary: Initiate Strava OAuth flow
      description: Redirects user to Strava authorization page to connect their account
      tags: [Authentication]
      responses:
        "302":
          description: Redirect to Strava OAuth page
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/strava/callback:
    get:
      summary: Handle Strava OAuth callback
      description: Processes the OAuth response from Strava and creates user session
      tags: [Authentication]
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
          description: Authorization code from Strava
        - in: query
          name: state
          required: true
          schema:
            type: string
          description: CSRF protection state token
      responses:
        "302":
          description: Redirect to frontend success or error page
...
  /api/auth/logout:
    post:
      summary: Logout user
      description: Destroys user session and clears authentication cookies
      tags: [Authentication]
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "500":
          description: Logout failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/revoke:
    delete:
      summary: Revoke Strava access and delete account
      description: Revokes access token with Strava and permanently deletes user account
      tags: [Authentication]
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Account deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # User Management Endpoints
...